rename : 코드 컨벤션에 맞추어 이름 변경
